everything is working now, but there is some optimization needed.

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: berminagovic <berminagovic@student.42.f    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/27 09:21:41 by bagovic           #+#    #+#             */
-/*   Updated: 2021/11/06 09:35:43 by berminagovi      ###   ########.fr       */
+/*   Updated: 2021/11/09 16:45:33 by berminagovi      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ void	ft_initialize_data(void);
 void	ft_putunsigned(long long n);
 int		ft_putnumber(long n);
 int		ft_putaddress(unsigned long long dec);
-int		ft_puthex(unsigned long long dec, int hexcase);
+int		ft_puthex(long dec, int hexcase);
 int		ft_printf(const char *format, ...);
 
 #endif
